# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "reeve",
    "description": "Fleet desired-state manager (Margo-inspired). Human API consumed by the reeve UI, plus the device-facing wire surface.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/devices": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "GET /api/devices (viewer+) — the fleet, ordered by device id.",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "All enrolled devices with presence, labels, layer-chain membership and current deployment states",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Below viewer role"
          }
        }
      }
    },
    "/api/devices/{device_id}": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "GET /api/devices/{device_id} (viewer+) — one device, with render\nprovenance.",
        "operationId": "detail",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "description": "Device id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device detail: summary plus render provenance (revision ids, manifest version, per-app secrets_version)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Below viewer role"
          },
          "404": {
            "description": "Unknown device",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{device_id}/journal": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "GET /api/devices/{device_id}/journal (viewer+) — status journal\npage, newest first (§7.3 forensic record; late-backfilled records\nappear in seq order, not arrival order).",
        "operationId": "journal",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "description": "Device id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size; default 100, max 1000",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "before_seq",
            "in": "query",
            "description": "Return records with seq strictly below this (older page)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One journal page, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalPage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Below viewer role"
          },
          "404": {
            "description": "Unknown device",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reeve/v1/manifest": {
      "get": {
        "tags": [
          "device"
        ],
        "summary": "GET /api/reeve/v1/manifest — the device-scoped State Manifest\n(§10.2). Renders on demand when the device's row is behind the\nlocal head (fresh enrollment, missed pass), then serves the stored\nmanifest bytes verbatim so the ETag is exact.",
        "operationId": "manifest",
        "parameters": [
          {
            "name": "if-none-match",
            "in": "header",
            "description": "RFC 9110 conditional GET against the manifest's strong ETag",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The device's current State Manifest (ETag header set)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateManifest"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match matched)"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "404": {
            "description": "Enrolled but never rendered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppManifestEntry": {
        "type": "object",
        "description": "Per-app State Manifest entry. The agent diffs `secrets_version`\nseparately from the bundle digest: bundle unchanged +\n`secrets_version` changed ⇒ re-resolve secrets and minimally\nre-up, no bundle re-pull (spec/reeve/10-secrets.md §12;\ndocs/decisions/agent.md).",
        "required": [
          "appId"
        ],
        "properties": {
          "appId": {
            "type": "string",
            "description": "Application id (matches the rendered app dir /\n`ApplicationDeployment.spec.applicationId`)."
          },
          "deploymentId": {
            "type": [
              "string",
              "null"
            ],
            "description": "Deterministic deployment id of the rendered\n`ApplicationDeployment` (docs/decisions/tree-render.md)."
          },
          "secrets_version": {
            "type": [
              "string",
              "null"
            ],
            "description": "Hash of resolved secret names+versions, never values\n(spec/reeve/10-secrets.md §12; docs/decisions/delivery.md D13).\nSpelled `secrets_version` on the wire — the exact token used\nnormatively throughout spec/reeve/."
          }
        }
      },
...
      "DeviceDeploymentState": {
        "type": "object",
        "description": "Current state of one deployment on a device\n(`deployment_status_current` row).",
        "required": [
          "deploymentId",
          "state",
          "receivedAt"
        ],
        "properties": {
          "deploymentId": {
            "type": "string"
          },
          "observedAt": {
            "type": [
              "string",
              "null"
            ],
            "description": "Device-assigned RFC 3339 timestamp (absent for vanilla reports)."
          },
          "receivedAt": {
            "type": "integer",
            "format": "int64",
            "description": "Server receipt time, unix seconds."
          },
          "seq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Journal seq of the report (absent for vanilla Margo reports)."
          },
          "state": {
            "type": "string",
            "description": "Margo deployment state (`pending` … `failed`)."
          }
        }
      },
      "DeviceDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DeviceSummary"
          },
          {
            "type": "object",
            "properties": {
              "render": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/RenderProvenance",
                    "description": "`null` when the device has never been rendered."
                  }
                ]
              }
            }
          }
        ],
        "description": "`GET /api/devices/{device_id}` response."
      },
...
      "DeviceSummary": {
        "type": "object",
        "description": "One device as listed by `GET /api/devices`.",
        "required": [
          "deviceId",
          "hostname",
          "arch",
          "agentVersion",
          "enrolledAt",
          "labels",
          "stale",
          "presence",
          "deployments"
        ],
        "properties": {
          "agentVersion": {
            "type": "string"
          },
          "arch": {
            "type": "string"
          },
          "class": {
            "type": [
              "string",
              "null"
            ],
            "description": "Layer-chain membership (D11: fleet -> class? -> region -> site\n-> device), each nullable."
          },
          "deployments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeviceDeploymentState"
            },
            "description": "Current per-deployment states."
          },
          "deviceId": {
            "type": "string"
          },
          "enrolledAt": {
            "type": "integer",
            "format": "int64",
            "description": "Unix seconds."
          },
          "hostname": {
            "type": "string"
          },
          "labels": {
            "type": "object",
            "description": "Free-form labels (docs/decisions/tree-render.md D12: labels\ngroup and filter, never configure).",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "lastSeenAt": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Unix seconds of last contact; `null` = never seen."
          },
          "presence": {
            "$ref": "#/components/schemas/PresenceInfo"
          },
          "region": {
            "type": [
              "string",
              "null"
            ]
          },
          "site": {
            "type": [
              "string",
              "null"
            ]
          },
          "stale": {
            "type": "boolean",
            "description": "Identity superseded by a newer enrollment from the same\nhostname (docs/decisions/agent.md D4 wiped-box case)."
          },
          "tierOrigin": {
            "type": [
              "string",
              "null"
            ],
            "description": "Child tier this device reached us through (federation §8.3);\n`null` = enrolled here."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The one error-body shape every JSON error response uses:\n`{ \"error\": \"<human-readable message>\" }`. Schema component for the\nD10 openapi.json (docs/decisions/ui.md); handlers keep building the\nbody via `json!` — this type documents the contract.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "JournalEntry": {
        "type": "object",
        "description": "One status-journal record (spec/reeve/05-health-journal.md §7.3).",
        "required": [
          "seq",
          "observedAt",
          "receivedAt",
          "kind"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "description": "`status` | `health` | `lifecycle` | `gap`."
          },
          "observedAt": {
            "type": "string",
            "description": "Device-assigned original RFC 3339 timestamp, verbatim."
          },
          "payload": {
            "description": "The journaled payload; absent for gap marks."
          },
          "receivedAt": {
            "type": "integer",
            "format": "int64",
            "description": "Server receipt time, unix seconds."
          },
          "seq": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "JournalPage": {
        "type": "object",
        "description": "`GET /api/devices/{device_id}/journal` response: newest first.",
        "required": [
          "records"
        ],
        "properties": {
          "nextBeforeSeq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Pass as `before_seq` to fetch the next (older) page; absent\nwhen this page reached the start of the journal."
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JournalEntry"
            }
          }
        }
      },
...
      "PresenceInfo": {
        "type": "object",
        "description": "Presence answer (spec/reeve/02-channel.md §4.3 vocabulary).",
        "required": [
          "state"
        ],
        "properties": {
          "since": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Unix seconds: channel-open time when online via channel,\nlast contact otherwise; `null` = never seen."
          },
          "state": {
            "$ref": "#/components/schemas/PresenceState",
            "description": "`online` | `offline`."
          }
        }
      },
...
      "RenderProvenance": {
        "type": "object",
        "description": "Render provenance of a device's current State Manifest\n(`device_manifests` row + the manifest's per-app entries;\ndocs/decisions/delivery.md D13).",
        "required": [
          "manifestVersion",
          "epoch",
          "counter",
          "generation",
          "contentDigest",
          "renderedRevision",
          "etag",
          "updatedAt",
          "apps"
        ],
        "properties": {
          "apps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppManifestEntry"
            },
            "description": "Per-app manifest entries — carries each app's `secrets_version`\n(spec/reeve/10-secrets.md §12.4)."
          },
          "bundleDigest": {
            "type": [
              "string",
              "null"
            ],
            "description": "OCI image manifest digest the device pulls; `null` = zero apps."
          },
          "contentDigest": {
            "type": "string",
            "description": "sha256 over the rendered apps/** file set (change detector)."
          },
          "counter": {
            "type": "integer",
            "format": "int64",
            "description": "Low 48 bits — monotonic per device.",
            "minimum": 0
          },
          "epoch": {
            "type": "integer",
            "format": "int32",
            "description": "High 16 bits — bumped by restore fencing\n(spec/reeve/07-durability.md §9.5).",
            "minimum": 0
          },
          "etag": {
            "type": "string",
            "description": "Strong ETag of the served manifest JSON."
          },
          "generation": {
            "type": "integer",
            "format": "int64",
            "description": "D2 render generation counter."
          },
          "manifestVersion": {
            "type": "integer",
            "format": "int64",
            "description": "The packed (epoch, counter) wire u64\n(spec/reeve/08-packaging.md §10.2).",
            "minimum": 0
          },
          "renderedRevision": {
            "type": "integer",
            "format": "int64",
            "description": "Local-stream revision id this device was last rendered against."
          },
          "updatedAt": {
            "type": "integer",
            "format": "int64",
            "description": "Unix seconds."
          }
        }
      },
...
      "StateManifest": {
        "type": "object",
        "description": "The reeve State Manifest body returned by\n`GET /api/reeve/v1/manifest` (spec/reeve/08-packaging.md §10.2) —\nState-Manifest-shaped after Margo's `UnsignedAppStateManifest`\n(`workload-management-api-1.0.0.yaml`): `manifestVersion`,\nrender-bundle digest + pull URL, per-app `secrets_version`.",
        "required": [
          "manifestVersion"
        ],
        "properties": {
          "apps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppManifestEntry"
            },
            "description": "Per-app entries carrying `secrets_version`\n(docs/decisions/delivery.md D13; spec/reeve/10-secrets.md §12)."
          },
          "bundle": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BundleRef",
                "description": "Render-bundle reference. Follows Margo's `DeploymentBundleRef`\nnull rule: with zero apps the property MUST be present with\nthe value `null`, never omitted — so it is always serialized."
              }
            ]
          },
          "manifestVersion": {
            "$ref": "#/components/schemas/ManifestVersion"
          }
        }
      },
...
    }
  },
...
}